Fix incorrect url resolution
The item's given_url and resolved_url are different in some
cases (e.g. the url with number sign). This is the problem in
server side, and it will result in the incorrect display of the icon, or
the error when the item is retrieved. This commit should fix it.